Make next/prev per-highway
Makes them dynamic instead of set - since next/prev sometimes need to be
in the context of a highway.

When not given a highway, it operates based on the first highway in the
list on the node, if available.

Since this eliminates the main purpose of update_links, rename that to
post_process, which is a stub in the superclass, but HwySeg still has
some link-finding to do.